Revert "move src/lib/platform.js to repo root"
This reverts commit 2738431b2bc825afa717f4ee34b1fef13131b76b.
