+++ src/lib/platform.test.js
@@ -1,4 +1,4 @@
-const { PLATFORMS, PlatformId, platformForId, isKnownPlatformId } = require('../../platform');
+const { PLATFORMS, PlatformId, platformForId, isKnownPlatformId } = require('./platform');
 const { expect } = require('../../test/setup');
 
 const deviceConstants = require('@particle/device-constants');
